calculated overall ratings

+++ Models/UserProfessorRating.cs
@@ -7,8 +7,8 @@ namespace MyRateApp2.Models
     [Keyless]
     public partial class UserProfessorRating
     {   
-        public long? UserId { get; set; }
-        public long? ProfRateId { get; set; }
+        public int? UserId { get; set; }
+        public int? ProfRateId { get; set; }
 
         public virtual ProfessorRating? ProfRate { get; set; }
         public virtual User? User { get; set; }
